Move the `jeff` generator header to `mlir/include`
