+++ mlir/benchmark/jeff/CMakeLists.txt
@@ -8,7 +8,6 @@
 
 # The lowering of a benchmark program to `jeff`
 add_library(MQTJeffBenchmarks STATIC Generate.cpp)
-target_include_directories(MQTJeffBenchmarks PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 target_link_libraries(MQTJeffBenchmarks PUBLIC MQTBenchmarkPrograms MQTCompilerPipeline)
 mqt_mlir_target_use_project_options(MQTJeffBenchmarks)
 llvm_update_compile_flags(MQTJeffBenchmarks)
