+++ mlir/unittests/benchmark/test_benchmark_jeff.cpp
@@ -8,7 +8,7 @@
  * Licensed under the MIT License
  */
 
-#include "Generate.h"
+#include "mlir/Benchmark/Jeff/Generate.h"
 #include "mlir/Benchmark/Programs.h"
 
 #include <gtest/gtest.h>
